Completed history page. Split up components in plan page. Redesign SelectTrip page
