+++ travel-planner/travel-planner-ui/src/Components/Banner/Banner.css
@@ -30,5 +30,4 @@
 	height: fit-content;
 	margin: auto;
 	text-shadow: 0 0 40px rgb(85, 38, 0);
-	cursor: default;
 }
