+++ travel-planner/travel-planner-ui/src/Components/Plan/Plan.css
@@ -1,84 +1,3 @@
 .plan-page {
 	overflow-y: auto;
 }
-
-.left-app {
-	flex: 2;
-}
-
-.right-app {
-	flex: 1;
-}
-
-.directions-mode-button {
-	background-color: rgba(85, 38, 0, 0.295);
-	border: rgb(85, 38, 0) solid 3px;
-	border-radius: 20px;
-	display: flex;
-	justify-content: space-between;
-	width: 14vw;
-	padding: 0px;
-	font-weight: bold;
-}
-
-.directions-mode-toggle {
-	background-color: rgb(85, 38, 0, 0.5);
-	border-radius: 0px 20px 20px 0px;
-	color: white;
-	font-weight: bolder;
-	transition: all 0.3s ease-in 0s;
-	flex: 1 !important;
-}
-
-.directions-mode-button span {
-	flex: 2;
-	padding: 0.6vh 0.4vw;
-}
-
-.directions-mode-toggle-on {
-	background-color: rgb(85, 38, 0);
-}
-
-.map-settings-panel {
-	display: flex;
-	justify-content: space-between;
-	padding-right: 2.6vw;
-	align-items: center;
-}
-
-.fa-box-archive {
-	font-size: 2.6vh;
-	padding-left: 2px;
-	transition: ease-in-out 0.8s;
-}
-
-.archive-button:hover .fa-box-archive {
-	font-size: 2.8vh;
-	transition: ease-in-out 0.8s;
-}
-
-.archive-button {
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	opacity: 34%;
-	transition: ease-in-out 0.8s;
-}
-
-.archive-button:hover {
-	opacity: 90%;
-	transition: ease-in-out 0.8s;
-}
-
-.archive-button > p {
-	letter-spacing: 2px;
-	font-size: 12px;
-	line-height: 0px;
-	transition: ease-in-out 0.8s;
-}
-
-.archive-button:hover > p {
-	letter-spacing: 3.6px;
-	font-size: 14px;
-	transition: ease-in-out 0.8s;
-}
