+++ travel-planner/travel-planner-ui/src/Components/SelectTrip/SelectTrip.css
@@ -1,106 +1,31 @@
-.create-new-trip-form {
-	display: flex;
-	justify-content: center;
+.trip-option {
+	cursor: pointer;
+	margin: 40px 0px;
 }
 
-.create-new-trip-form form {
-	box-shadow: 0 6px 15px 0 rgb(85, 38, 0);
-	background-color: bisque;
-	position: absolute;
-	z-index: 10;
-	width: 86vw;
-	height: 60vh;
-	border-radius: 20px;
-	padding: 6vh 4vw;
-	margin-top: 3.6vh;
+.select-trip-page-top {
 	display: flex;
 	align-items: center;
-	justify-content: space-evenly;
-	flex-direction: column;
+	justify-content: space-between;
 }
 
-.create-new-trip-form input {
-	margin: 0.7vh 0px;
-	padding: 0.5vh 1.3vw;
-	width: 100%;
-	border-radius: 20px;
-	border: none;
-	height: 3vh;
-	text-overflow: clip;
-	letter-spacing: 0.2vh;
-	font-size: 1.6vh;
-	flex: 14;
+.select-trip-page-top > h1 {
+	width: auto;
 }
 
-.create-new-trip-form label {
-	display: flex;
-	align-items: center;
-	width: 100%;
-}
-
-.create-new-trip-form h3 {
-	margin-right: 2vw;
-	font-size: 1.6vh;
-	letter-spacing: 0.2vh;
+.trip-search {
 	flex: 1;
 }
 
-.create-new-trip-form button {
-	margin-bottom: 4vh;
-
-	text-overflow: clip;
-	font-size: 2vh;
-	letter-spacing: 0.5vw;
-	margin-top: 3vh;
+.trip-search input {
+	width: 90%;
+	height: 2.6vh;
+	background-color: bisque;
+	outline: none;
+	border: 1px rgba(85, 38, 0, 0.39) solid;
 	border-radius: 20px;
-	width: 54%;
-	font-weight: bold;
-	color: white;
-	cursor: pointer;
-	height: 5vh;
-	border: none;
-	background-size: 300% 100%;
-
-	-o-transition: all 0.7s ease-in-out;
-	-webkit-transition: all 0.7s ease-in-out;
-	transition: all 0.7s ease-in-out;
-
-	background-image: linear-gradient(
-		to right,
-		rgb(255, 249, 168),
-		rgb(223, 168, 120),
-		rgb(223, 168, 120),
-		rgb(255, 219, 122)
-	);
-	box-shadow: 0 4px 15px 0 rgba(85, 38, 0, 0.459);
-}
-
-.create-new-trip-form button:hover {
-	background-position: 100% 0;
-	-o-transition: all 0.7s ease-in-out;
-	-webkit-transition: all 0.7s ease-in-out;
-	transition: all 0.7s ease-in-out;
-}
-
-.create-new-trip-form div {
-	width: 100%;
-}
-
-.dropdown > a {
-	font-size: 2vh;
-	padding: 1vh 0px;
-}
-
-.dropdown a {
-	color: rgb(85, 38, 0);
-	padding: 0.5vh 0px;
-}
-
-.dropdown a:hover {
-	font-weight: bold;
-}
-
-.list-unstyled {
-	display: flex;
-	flex-direction: column;
+	padding: 0px 1vw;
+	box-shadow: 0px 11px 35px 2px rgba(85, 38, 0, 0.329);
+	transition: all 0.5s ease-in-out;
+	margin: 0px 3.4vw;
 }
