+++ travel-planner/travel-planner-ui/src/Components/SelectTrip/SelectTrip.jsx
@@ -1,127 +1,109 @@
 import "./SelectTrip.css";
-import React, { useState, useEffect, useContext } from "react";
-import axios from "axios";
-import { UserContext } from "../../UserContext";
-import Dropdown from "react-bootstrap/Dropdown";
-import CustomToggle from "./../Dropdown/CustomToggle";
-import CustomMenu from "./../Dropdown/CustomMenu";
-import Loading from "../Loading/Loading";
 import { PORT } from "./../App/App";
+import { UserContext } from "../../UserContext";
+import Loading from "../Loading/Loading";
+import Banner from "../Banner/Banner";
+import { useEffect, useContext, useState } from "react";
+import axios from "axios";
 
-function CreateNewTrip(props) {
-	const [newTripTitle, setNewTripTitle] = useState(props.tripTitle);
-	const [newTripCenter, setNewTripCenter] = useState("");
-
-	const submitNewTrip = async (event) => {
-		event.preventDefault();
-
-		// Create a new trip
-		// Then send that newly create trip back to Plans through setFinalTripDetails
-
-		const res = await axios.post(`http://localhost:${PORT}/maps/newMap`, {
-			title: newTripTitle,
-			center: newTripCenter,
-			sessionToken: props.sessionToken,
-		});
-		props.setFinalTripDetails(res.data);
-	};
-
-	return (
-		<div className="create-new-trip-form">
-			<form onSubmit={submitNewTrip}>
-				<h1>Create a new trip!</h1>
-				<div>
-					<label>
-						<h3>Title</h3>
-						<input
-							type="text"
-							value={newTripTitle}
-							onChange={(event) => setNewTripTitle(event.currentTarget.value)}
-							className="new-trip-title"
-							required
-						></input>
-					</label>
-					<label>
-						<h3>General Location</h3>
-						<input
-							type="text"
-							placeholder="i.e. city, state, country"
-							value={newTripCenter}
-							onChange={(event) => {
-								setNewTripCenter(event.currentTarget.value);
-							}}
-						></input>
-					</label>
-				</div>
-				<button type="submit">Create trip</button>
-			</form>
-		</div>
-	);
-}
-
-export default function SelectTripPage(props) {
+export default function SelectTrip({
+	setTripDetails,
+	isArchived,
+	title,
+	onEnter,
+}) {
 	const { sessionToken } = useContext(UserContext);
+	const [trips, setTrips] = useState({
+		term: "",
+		filteredTrips: [],
+		allTrips: null,
+	});
 
-	const [creatingTrip, setCreatingTrip] = useState("");
-	const [trips, setTrips] = useState(null); // An array containing Map objects for the user
-
-	// Get the existing trips by querying
+	// Get archived trips from database
 	useEffect(async () => {
 		const res = await axios.post(`http://localhost:${PORT}/maps/getUserMaps`, {
 			sessionToken: sessionToken,
-			isArchived: false,
+			isArchived: isArchived,
 		});
 
-		setTrips(res.data);
+		setTrips((prev) => ({
+			...prev,
+			allTrips: res.data,
+			filteredTrips: res.data,
+		}));
 	}, []);
+
+	if (!trips) {
+		return <Loading />;
+	}
 
 	const onClickTripName = (event) => {
 		const tripId = event.currentTarget.getAttribute("data-objectid");
 
 		trips.forEach((trip) => {
 			if (trip.objectId === tripId) {
-				props.setTripDetails(trip);
+				setTripDetails(trip);
 			}
 		});
 	};
 
-	if (!trips) {
-		return <Loading />;
-	}
+	const updateFilteredTrips = (newSearchTerm) => {
+		let filteredTrips = [];
+
+		if (newSearchTerm !== "") {
+			trips.allTrips.forEach((trip) => {
+				if (trip.MapName.toLowerCase().includes(newSearchTerm.toLowerCase())) {
+					filteredTrips.push(trip);
+				}
+			});
+		} else {
+			filteredTrips = trips.allTrips;
+		}
+
+		setTrips((prev) => ({
+			...prev,
+			term: newSearchTerm,
+			filteredTrips: filteredTrips,
+		}));
+	};
 
 	return (
 		<div className="select-trip-page">
-			<Dropdown>
-				<Dropdown.Toggle as={CustomToggle} id="dropdown-custom-components">
-					Select your trip!
-				</Dropdown.Toggle>
+			<div className="select-trip-page-top">
+				<h1>{title}</h1>
+				<div className="trip-search">
+					<input
+						onChange={(event) => updateFilteredTrips(event.target.value)}
+						onKeyDown={
+							onEnter
+								? (event) => {
+										if (event.key === "Enter") {
+											onEnter(trips.term);
+										}
+								  }
+								: null
+						}
+						value={trips.term}
+						placeholder="Search..."
+					></input>
+				</div>
+			</div>
+			{trips.length === 0 && <h2>No trips found.</h2>}
 
-				<Dropdown.Menu
-					className="outer-dropdown"
-					onKeyDown={(event) => {
-						if (event.key === "Enter")
-							setCreatingTrip(event.currentTarget.value);
-					}}
-					as={CustomMenu}
+			{trips.filteredTrips.map((trip) => (
+				<div
+					key={trip.mapName}
+					data-objectid={trip.objectId}
+					onClick={onClickTripName}
+					className="trip-option"
 				>
-					{trips.map((trip) => (
-						<Dropdown.Item
-							key={trip.objectId}
-							data-objectid={trip.objectId}
-							onClick={onClickTripName}
-						>
-							{trip.MapName}
-						</Dropdown.Item>
-					))}
-				</Dropdown.Menu>
-			</Dropdown>
-			{creatingTrip != "" && (
-				<CreateNewTrip
-					tripTitle={creatingTrip}
-					setFinalTripDetails={props.setTripDetails}
-					sessionToken={sessionToken}
-				/>
-			)}
+					<Banner
+						text={trip.MapName}
+						cursorStyle={"pointer"}
+						imgSrc="https://images.unsplash.com/photo-1507525428034-b723cf961d3e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8dHJvcGljYWwlMjBiZWFjaHxlbnwwfHwwfHw%3D&w=1000&q=80"
+					/>
+				</div>
+			))}
 		</div>
 	);
 }
